requirements.txt
telethon
supabase
python-dotenv
pytz
httpx
